Reach, Optimizing and plenty of bug fixes
- Fixing Aim (E) false positives.
- Fixing Aim (B) not flagging.
- Optimized check processing.
- Optimized alerts and debugging.
- Optimized DataManager
- Fixing concurrency problems on shutdown.
- Working on Reach (B) more
- Added instant actions optional flush or queue.
- Fixing Aim (B) not flagging as much as it should.
- Fixing NoFall (B) SPOOF_AIR false positive.
- Optimized Aim (J).
- Fixing Aim (I) false positive.
-

+++ Impl/src/main/java/dev/brighten/anticheat/data/classes/PlayerInformation.java
@@ -75,7 +75,7 @@ public class PlayerInformation {
     public boolean cinematicMode;
 
     //Gcd
-    public int yawGCD, pitchGCD, lastYawGCD, lastPitchGCD;
+    public float yawGCD, pitchGCD, lastYawGCD, lastPitchGCD;
 
     //Server Position
     public long lastServerPos, lastRespawn;
